Show a warning if a select is required but the search is not enabled.
The required validation error is attached to an input (the search box),
so if there is no search box we cannot use setCustomValidity() to show
the HTML5 validation error.
Remove the required attribute from the original select to prevent the
browser from blocking form submission and displaying a console error.